Drop Django 1.6 support and reorganize code.
Move `models.py` to `handlers.py`.
Add back the original `pre_save` as a fallback.
Make the cache into its own module in `cache.py` with related functions.
Make a public module in `cleanup.py` to manually refresh the instance cache.
Workaround a bug in `refresh_from_db` which leaves FileFields with reference
to the wrong instance. (Reported: khamaileon)
Workaround a bug in the FileFields descriptor which doesn't recover Fields or
storage classes after a pickle/unpickle. (Reported: khamaileon)
Workaround behavior of a FieldFile's delete method which sets the
field to `None` on the instance. (Reported: Pierre Chiquet)
Add tests for all of these workarounds.